# apps/docs/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v3/auth/token": {
      "post": {
        "operationId": "Auth_ExchangeToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token successfully exchanged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Exchange client credentials for an access token",
        "tags": [
          "V3 Auth"
        ]
      }
    },
    "/v3/auth/oauth2/{path}": {
      "get": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_get",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "post": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_post",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "put": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_put",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "delete": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_delete",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "patch": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_patch",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "options": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_options",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "head": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_head",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      },
      "search": {
        "description": "Proxies OAuth2 requests (authorize, token, userinfo, etc.) to the auth provider",
        "operationId": "AuthController_handleOAuth2_search",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Handle OAuth2 provider requests",
        "tags": [
          "V3 Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Scryme V3 API",
    "description": "The Scryme V3 API documentation - Scalable & Enterprise Ready",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "example": {
              "message": "Invalid request parameters",
              "code": "BAD_REQUEST",
              "details": [
                "email must be an email"
              ]
            }
          },
          "timestamp": {
            "type": "string",
            "example": "2023-10-27T10:00:00.000Z"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp"
        ]
      },
...
      "TokenRequestDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "example": "client_id_123",
            "description": "The Client ID provided during provisioning or registration"
          },
          "clientSecret": {
            "type": "string",
            "example": "client_secret_456",
            "description": "The Client Secret"
          }
        },
        "required": [
          "clientId",
          "clientSecret"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "example": 3600
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn"
        ]
      },
...
    }
  }
}